+++ src/app/user/[id]/setting/edit/components/EditBodyInfo/EditBodyInfo.tsx
@@ -3,8 +3,8 @@ import * as S from "./EditBodyInfo.styles";
 import { FieldErrors, UseFormRegister } from "react-hook-form";
 import { UserEditData } from "../../UserEdit.types";
 import { USER_EDIT_PLACEHOLDER } from "../../UserEdit.constants";
-import { UserEditInput } from "..";
 import { validation_user } from "@/lib/constants/userValidate";
+import { UserEditInput } from "@/components";
 
 interface EditBodyInfoProps {
   register: UseFormRegister<UserEditData>;
