+++ src/app/user/[id]/setting/edit/components/index.ts
@@ -3,5 +3,3 @@ export { default as EditSex } from "./EditSex/EditSex";
 export { default as EditBirthDate } from "./EditBirthDate/EditBirthDate";
 export { default as EditBodyInfo } from "./EditBodyInfo/EditBodyInfo";
 export { default as EditIntensity } from "./EditIntensity/EditIntensity";
-
-export { default as UserEditInput } from "./UserEditInput/UserEditInput";
