+++ src/components/index.ts
@@ -14,3 +14,4 @@ export { default as PinDetailCard } from "./PinDetailCard";
 export { default as Window } from "./Window";
 export { default as Textarea } from "./Textarea";
 export { default as InputRadio } from "./InputRadio/InputRadio";
+export { default as UserEditInput } from "./UserEditInput/UserEditInput";
